refactor(v1): organize NeMo Gym adapter
