+++ verifiers/v1/mcp/launch.py
@@ -17,8 +17,6 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 from urllib.parse import parse_qsl, urlencode, urlsplit, urlunsplit
 
-import httpx
-
 from verifiers.v1.errors import ToolsetError
 from verifiers.v1.interception.tunnel import PrimeTunnel
 from verifiers.v1.mcp.server import (
@@ -37,12 +35,9 @@ from verifiers.v1.runtimes.base import _ENSURE_UV
 from verifiers.v1.state import State
 
 if TYPE_CHECKING:
-    from mcp import ClientSession
-
     from verifiers.v1.mcp.toolset import Toolset
 
 logger = logging.getLogger(__name__)
-_MCP_TIMEOUT = httpx.Timeout(600.0, connect=5.0)
 
 # Sandboxed servers install the working tree, so only wheel inputs need to cross the boundary.
 VF_BUILD_INPUTS = ("pyproject.toml", "README.md", "LICENSE", "verifiers")
@@ -498,36 +493,3 @@ async def serve_tools(
                 )
                 logger.info("tool server '%s': %s", name, urls[name])
         yield urls
-
-
-@contextlib.asynccontextmanager
-async def mcp_session(
-    url: str,
-    *,
-    headers: dict[str, str] | None = None,
-    timeout: float | httpx.Timeout = _MCP_TIMEOUT,
-) -> AsyncIterator[ClientSession]:
-    """One fresh session to an MCP server, opened and closed within the caller's task so AnyIO
-    cancellation scopes stay correctly nested. A teardown failure after the body completed is
-    suppressed — the result is already in hand, and closing noise must not fail (or replay) an
-    already-answered call."""
-    from mcp import ClientSession
-    from mcp.client.streamable_http import (
-        create_mcp_http_client,
-        streamable_http_client,
-    )
-
-    stack = contextlib.AsyncExitStack()
-    try:
-        http_client = await stack.enter_async_context(
-            create_mcp_http_client(headers=headers, timeout=timeout)
-        )
-        read, write, *_ = await stack.enter_async_context(
-            streamable_http_client(url, http_client=http_client)
-        )
-        session = await stack.enter_async_context(ClientSession(read, write))
-        await session.initialize()
-        yield session
-    finally:
-        with contextlib.suppress(Exception):
-            await stack.aclose()
