+++ verifiers/v1/tasksets/nemo_gym/__init__.py
@@ -2,11 +2,11 @@ from verifiers.v1.tasksets.nemo_gym.taskset import (
     NeMoGymConfig,
     NeMoGymData,
     NeMoGymEnv,
-    NeMoGymState,
     NeMoGymTask,
     NeMoGymTaskConfig,
     NeMoGymTaskset,
 )
+from verifiers.v1.tasksets.nemo_gym.toolset import NeMoGymState, NeMoGymToolset
 
 __all__ = [
     "NeMoGymConfig",
@@ -16,4 +16,5 @@ __all__ = [
     "NeMoGymTask",
     "NeMoGymTaskConfig",
     "NeMoGymTaskset",
+    "NeMoGymToolset",
 ]
